requirements.txt
asgiref==3.3.4
Django==3.1.7
django-braces==1.14.0
django-debug-toolbar==3.2
django-extensions==3.1.2
djangorestframework==3.12.4
install==1.3.4
jsonfield==3.1.0
model-bakery==1.3.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
urllib3==1.26.4